+++ packages/save-and-share-bar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.116](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.9.115...@times-components/save-and-share-bar@0.9.116) (2022-06-22)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
 ## [0.9.115](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.9.114...@times-components/save-and-share-bar@0.9.115) (2022-06-08)
 
 **Note:** Version bump only for package @times-components/save-and-share-bar